refactor(themes): move themes to /utils/themes
