+++ src/components/react_canvas/axis.tsx
@@ -11,7 +11,7 @@ import {
   isVertical,
 } from '../../lib/axes/axis_utils';
 import { AxisSpec, Position } from '../../lib/series/specs';
-import { Theme } from '../../lib/themes/theme';
+import { Theme } from '../../utils/themes/theme';
 import { Dimensions } from '../../utils/dimensions';
 
 interface AxisProps {
