+++ src/components/react_canvas/bar_values.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Group, Rect, Text } from 'react-konva';
 import { BarGeometry } from '../../lib/series/rendering';
 import { Rotation } from '../../lib/series/specs';
-import { DisplayValueStyle } from '../../lib/themes/theme';
+import { DisplayValueStyle } from '../../utils/themes/theme';
 import { Dimensions } from '../../utils/dimensions';
 import { buildBarValueProps } from './utils/rendering_props_utils';
