+++ src/components/react_canvas/grid.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Group, Line } from 'react-konva';
 import { AxisLinePosition } from '../../lib/axes/axis_utils';
-import { DEFAULT_GRID_LINE_CONFIG, GridLineConfig, mergeWithDefaultGridLineConfig } from '../../lib/themes/theme';
+import { DEFAULT_GRID_LINE_CONFIG, GridLineConfig, mergeWithDefaultGridLineConfig } from '../../utils/themes/theme';
 import { Dimensions } from '../../utils/dimensions';
 
 interface GridProps {
