+++ src/components/react_canvas/line_geometries.tsx
@@ -4,7 +4,7 @@ import { Circle, Group, Path } from 'react-konva';
 import { animated, Spring } from 'react-spring/renderprops-konva.cjs';
 import { LegendItem } from '../../lib/series/legend';
 import { getGeometryStyle, LineGeometry, PointGeometry } from '../../lib/series/rendering';
-import { SharedGeometryStyle } from '../../lib/themes/theme';
+import { SharedGeometryStyle } from '../../utils/themes/theme';
 import {
   buildLineRenderProps,
   buildPointStyleProps,
