+++ src/components/react_canvas/reactive_chart.tsx
@@ -3,7 +3,7 @@ import { inject, observer } from 'mobx-react';
 import React from 'react';
 import { Layer, Rect, Stage } from 'react-konva';
 import { isLineAnnotation, isRectAnnotation, Position } from '../../lib/series/specs';
-import { LineAnnotationStyle, RectAnnotationStyle } from '../../lib/themes/theme';
+import { LineAnnotationStyle, RectAnnotationStyle } from '../../utils/themes/theme';
 import { AnnotationId } from '../../utils/ids';
 import { AnnotationDimensions, AnnotationLineProps, AnnotationRectProps } from '../../state/annotation_utils';
 import { ChartStore, Point } from '../../state/chart_state';
