+++ src/components/react_canvas/rect_annotation.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Group, Rect } from 'react-konva';
-import { RectAnnotationStyle } from '../../lib/themes/theme';
+import { RectAnnotationStyle } from '../../utils/themes/theme';
 import { Dimensions } from '../../utils/dimensions';
 import { AnnotationRectProps } from '../../state/annotation_utils';
