+++ src/lib/axes/axis_utils.test.ts
@@ -1,7 +1,7 @@
 import { XDomain } from '../series/domains/x_domain';
 import { YDomain } from '../series/domains/y_domain';
 import { AxisSpec, DomainRange, Position, AxisStyle } from '../series/specs';
-import { LIGHT_THEME } from '../themes/light_theme';
+import { LIGHT_THEME } from '../../utils/themes/light_theme';
 import { AxisId, getAxisId, getGroupId, GroupId } from '../../utils/ids';
 import { ScaleType } from '../../utils/scales/scales';
 import {
