+++ src/lib/series/rendering.bars.test.ts
@@ -5,7 +5,7 @@ import { ScaleType } from '../../utils/scales/scales';
 import { renderBars } from './rendering';
 import { computeXScale, computeYScales } from './scales';
 import { BarSeriesSpec } from './specs';
-import { LIGHT_THEME } from '../themes/light_theme';
+import { LIGHT_THEME } from '../../utils/themes/light_theme';
 
 const SPEC_ID = getSpecId('spec_1');
 const GROUP_ID = getGroupId('group_1');
