+++ src/lib/series/rendering.lines.test.ts
@@ -7,7 +7,7 @@ import { CurveType } from '../../utils/curves';
 import { IndexedGeometry, LineGeometry, PointGeometry, renderLine } from './rendering';
 import { computeXScale, computeYScales } from './scales';
 import { LineSeriesSpec } from './specs';
-import { LIGHT_THEME } from '../themes/light_theme';
+import { LIGHT_THEME } from '../../utils/themes/light_theme';
 const SPEC_ID = getSpecId('spec_1');
 const GROUP_ID = getGroupId('group_1');
