+++ src/lib/series/rendering.test.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_GEOMETRY_STYLES } from '../themes/theme_commons';
+import { DEFAULT_GEOMETRY_STYLES } from '../../utils/themes/theme_commons';
 import { getSpecId } from '../../utils/ids';
 import { BarGeometry, getGeometryStyle, isPointOnGeometry, PointGeometry } from './rendering';
