+++ src/lib/series/series.test.ts
@@ -1,4 +1,4 @@
-import { ColorConfig } from '../themes/theme';
+import { ColorConfig } from '../../utils/themes/theme';
 import { getGroupId, getSpecId, SpecId } from '../../utils/ids';
 import { ScaleType } from '../../utils/scales/scales';
 import {
