+++ src/lib/series/series.ts
@@ -1,4 +1,4 @@
-import { ColorConfig } from '../themes/theme';
+import { ColorConfig } from '../../utils/themes/theme';
 import { Accessor } from '../../utils/accessor';
 import { GroupId, SpecId } from '../../utils/ids';
 import { splitSpecsByGroupId, YBasicSeriesSpec } from './domains/y_domain';
