+++ src/lib/utils/interactions.test.ts
@@ -1,5 +1,4 @@
 import { IndexedGeometry, PointGeometry } from '../series/rendering';
-import { Dimensions } from './dimensions';
 import { getSpecId } from '../../utils/ids';
 import {
   areIndexedGeometryArraysEquals,
@@ -10,6 +9,7 @@ import {
   isFollowTooltipType,
   TooltipType,
 } from './interactions';
+import { Dimensions } from '../../utils/dimensions';
 
 const seriesStyle = {
   rect: {
