+++ src/specs/line_annotation.tsx
@@ -1,7 +1,7 @@
 import { inject } from 'mobx-react';
 import React, { createRef, CSSProperties, PureComponent } from 'react';
 import { LineAnnotationSpec } from '../lib/series/specs';
-import { DEFAULT_ANNOTATION_LINE_STYLE } from '../lib/themes/theme';
+import { DEFAULT_ANNOTATION_LINE_STYLE } from '../utils/themes/theme';
 import { getGroupId } from '../utils/ids';
 import { SpecProps } from './specs_parser';
