+++ src/specs/settings.tsx
@@ -1,9 +1,9 @@
 import { inject } from 'mobx-react';
 import { PureComponent } from 'react';
 import { DomainRange, Position, Rendering, Rotation } from '../lib/series/specs';
-import { DARK_THEME } from '../lib/themes/dark_theme';
-import { LIGHT_THEME } from '../lib/themes/light_theme';
-import { BaseThemeType, mergeWithDefaultTheme, PartialTheme, Theme, BaseThemeTypes } from '../lib/themes/theme';
+import { LIGHT_THEME } from '../utils/themes/light_theme';
+import { DARK_THEME } from '../utils/themes/dark_theme';
+import { BaseThemeType, mergeWithDefaultTheme, PartialTheme, Theme, BaseThemeTypes } from '../utils/themes/theme';
 import { Domain } from '../utils/domain';
 import { TooltipType, TooltipValueFormatter } from '../lib/utils/interactions';
 import {
