+++ src/state/annotation_marker.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 import { AnnotationDomainType, AnnotationDomainTypes, AnnotationSpec, Position, Rotation } from '../lib/series/specs';
-import { DEFAULT_ANNOTATION_LINE_STYLE } from '../lib/themes/theme';
+import { DEFAULT_ANNOTATION_LINE_STYLE } from '../utils/themes/theme';
 import { Dimensions } from '../utils/dimensions';
 import { getAnnotationId, getGroupId, GroupId } from '../utils/ids';
 import { ScaleContinuous } from '../utils/scales/scale_continuous';
