+++ src/state/chart_state.timescales.test.ts
@@ -1,7 +1,7 @@
 import { DateTime } from 'luxon';
 import { LineSeriesSpec } from '../lib/series/specs';
-import { LIGHT_THEME } from '../lib/themes/light_theme';
-import { mergeWithDefaultTheme } from '../lib/themes/theme';
+import { LIGHT_THEME } from '../utils/themes/light_theme';
+import { mergeWithDefaultTheme } from '../utils/themes/theme';
 import { getGroupId, getSpecId } from '../utils/ids';
 import { ScaleType } from '../utils/scales/scales';
 import { ChartStore } from './chart_state';
