+++ src/state/utils.test.ts
@@ -11,7 +11,7 @@ import {
   LineSeriesSpec,
 } from '../lib/series/specs';
 import { BARCHART_1Y0G, BARCHART_1Y1G } from '../utils/data_samples/test_dataset';
-import { LIGHT_THEME } from '../lib/themes/light_theme';
+import { LIGHT_THEME } from '../utils/themes/light_theme';
 import { AxisId, getGroupId, getSpecId, SpecId } from '../utils/ids';
 import { ScaleContinuous } from '../utils/scales/scale_continuous';
 import { ScaleType } from '../utils/scales/scales';
