+++ stories/styling.tsx
@@ -23,7 +23,7 @@ import {
   LineSeriesStyle,
 } from '../src/';
 import * as TestDatasets from '../src/utils/data_samples/test_dataset';
-import { palettes } from '../src/lib/themes/colors';
+import { palettes } from '../src/utils/themes/colors';
 
 function range(title: string, min: number, max: number, value: number, groupId?: string, step: number = 1) {
   return number(
